Unified health and healthcheck files
