+++ go.mod
@@ -8,5 +8,6 @@ require (
 	github.com/globalsign/mgo v0.0.0-20181015135952-eeefdecb41b8
 	github.com/kr/pretty v0.2.0 // indirect
 	github.com/smartystreets/goconvey v1.6.4
+	golang.org/x/net v0.0.0-20190404232315-eb5bcb51f2a3 // indirect
 	gopkg.in/check.v1 v1.0.0-20190902080502-41f04d3bba15 // indirect
 )
